Update find server versions
