+++ lib/BdsSystemInfo.js
@@ -4,11 +4,16 @@ const { execSync } = require("child_process");
 const commadExist = require("./commandExist");
 const Request = require("./Requests");
 const bds = require("../index");
+const { PlatformVersionsV2 } = require("../src/BdsServersDownload");
 
 async function CheckSystemAsync() {
     const
         PHPBin = await Request.JSON("https://raw.githubusercontent.com/The-Bds-Maneger/Php_Static_Binary/main/binarys.json"),
-        Servers = await Request.JSON("https://raw.githubusercontent.com/The-Bds-Maneger/external_files/main/Server.json");
+        Servers = {
+            bedrock: await PlatformVersionsV2("bedrock"),
+            spigot: await PlatformVersionsV2("spigot"),
+            dragonfly: await PlatformVersionsV2("dragonfly"),
+        }
     
     let system, require_qemu = false;
     const valid_platform = {
@@ -23,7 +28,7 @@ async function CheckSystemAsync() {
     else valid_platform["pocketmine"] = false;
 
     // 
-    if (!(Servers.dragonfly[Servers.latest.dragonfly][process.platform][bds.arch])) valid_platform["dragonfly"] = false;
+    if (!(Servers.dragonfly.versions[Servers.dragonfly.latest][process.platform][bds.arch])) valid_platform["dragonfly"] = false;
 
     // SoSystem X
     if (process.platform == "win32") {
@@ -32,8 +37,8 @@ async function CheckSystemAsync() {
         system = "Linux";
         
         // Bedrock Check
-        if (Servers.bedrock[Servers.latest.bedrock][arch]) {
-            if (Servers.bedrock[Servers.latest.bedrock][arch][process.platform]) valid_platform["bedrock"] = true;
+        if (Servers.bedrock.versions[Servers.bedrock.latest][process.platform]) {
+            if (Servers.bedrock.versions[Servers.bedrock.latest][process.platform][arch]) valid_platform["bedrock"] = true;
             else valid_platform["bedrock"] = false;
         } else valid_platform["bedrock"] = false;
 
@@ -46,22 +51,21 @@ async function CheckSystemAsync() {
         }
     } else if (process.platform == "darwin") {
         system = "MacOS";
-        valid_platform["bedrock"] = false
     } else if (process.platform === "android") {
         system = "Android";
-        valid_platform["bedrock"] = false
     } else {
         throw new Error(`The Bds Maneger Core does not support ${process.platform} systems, as no tests have been done.`);
     }
 
     return {
-        system,
-        require_qemu,
-        valid_platform,
+        system: system,
+        require_qemu: require_qemu,
+        valid_platform: valid_platform,
     };
 }
 
 module.exports = CheckSystemAsync;
+module.exports.CheckSystemAsync = CheckSystemAsync;
 
 // System Architect (x64, aarch64 and others)
 let arch;
